+++ lib/glimmer/swt/custom/shape/polygon.rb
@@ -156,8 +156,12 @@ module Glimmer
           end
 
           def include?(x, y)
-            comparison_lines = absolute_point_xy_array.zip(absolute_point_xy_array.rotate(1))
-            comparison_lines.any? {|line| Line.include?(line.first.first, line.first.last, line.last.first, line.last.last, x, y)}
+            if filled?
+              contain?(x, y)
+            else
+              comparison_lines = absolute_point_xy_array.zip(absolute_point_xy_array.rotate(1))
+              comparison_lines.any? {|line| Line.include?(line.first.first, line.first.last, line.last.first, line.last.last, x, y)}
+            end
           end
                     
           def move_by(x_delta, y_delta)