+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 4.18.7.3
 
+- Support the ability for nested shapes to override their parent `shape` common shared properties
 - Refactor Tetris to use a custom shape (`bevel`) for its blocks given they are used in both the game and the icon, thus achieving code reuse
 - Fix issue with moving filled polygon (moving drawn polygon works)